Dev version 0.0.12

+++ creep-tenuous-common/src/main/java/com/zer0s2m/creeptenuous/common/query/IMapFileObjectToCategory.java
@@ -12,7 +12,7 @@ public interface IMapFileObjectToCategory {
      * Get user ID entity
      * @return ID
      */
-    Integer getUserCategoryId();
+    Long getUserCategoryId();
 
     /**
      * Get system names of file objects
